requirements.txt
discord.py==1.7.3
googlesearch_python==1.1.0
python-dotenv==0.20.0
requests
beautifulsoup4==4.9.3
google==3.0.0
numpy==1.22.3
